python/requirements.txt
Flask==3.0.3
google-generativeai==0.7.0
python-dotenv==1.0.1
flask-limiter==3.5.0
pymongo==4.8.0
dnspython==2.6.1
gunicorn==22.0.0
apscheduler==3.10.4
requests==2.32.3
pydantic==2.8.2
sympy==1.13.2